backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pyjwt
passlib[bcrypt]
pydantic
email-validator
python-multipart
